requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2==2.9.3
drf-nested-routers==0.93.4
django-filter==22.1